requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.6
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
idna==2.10
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
